requirements.txt
# Core dependencies
pyyaml==6.0.1
bioc==2.1
langchain==0.3.25
langchain-openai==0.3.18
langchain-together==0.3.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Visualization and analysis
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.1 